requirements.txt
# AppOS — Python Low-Code Platform
# Core dependencies

# Web framework
reflex>=0.6.0

# Database
sqlalchemy[asyncio]>=2.0
alembic>=1.13
psycopg2-binary>=2.9

# Validation
pydantic>=2.5

# Cache & Sessions
redis>=5.0

# Task queue
celery[redis]>=5.3

# Dependency graph
networkx>=3.2

# Security
bcrypt>=4.1
cryptography>=42.0
passlib>=1.7

# Configuration
pyyaml>=6.0

# HTTP client (for integrations)
httpx>=0.27
